Split evaluator Value enum into separate file
